Split MqttManager tests and fix static analysis findings
