+++ catroid/src/main/java/org/catrobat/catroid/content/ActionFactory.java
@@ -105,8 +105,8 @@ import org.catrobat.catroid.content.actions.PlayNoteForBeatsAction;
 import org.catrobat.catroid.content.actions.PlaySoundAction;
 import org.catrobat.catroid.content.actions.PlaySoundAtAction;
 import org.catrobat.catroid.content.actions.PointInDirectionAction;
-import org.catrobat.catroid.content.actions.PublishMqttMessageAction;
 import org.catrobat.catroid.content.actions.PointToAction;
+import org.catrobat.catroid.content.actions.PublishMqttMessageAction;
 import org.catrobat.catroid.content.actions.RaspiIfLogicAction;
 import org.catrobat.catroid.content.actions.RaspiPwmAction;
 import org.catrobat.catroid.content.actions.RaspiSendDigitalValueAction;
